feat: add dependency visualization
Add visualization of the dependencies between ingredients, equipment, and actions.

This includes the key functionality of the visualization, as it seems that we started skipping a number of dependencies somewhere along the way (the Actions show quite a bit of duplication, when examined more closely).

The visualization should be useful from development/debugging perspective, and the skipped dependencies are pre-existing, so we will proceed with merging this.