refactor: identify Actions earlier
Change to _identifying_ the Actions in the first parsing pass, so that the second pass is more focused. The _dependencies/relationships_ continue to be identified in the second pass, as before.

The split allows us leverage the powerful flexibility of the plain LLM request/"first pass", which is free to leverage its semantic knowledge to identify the actions, and is not constrained to the tools we provide as the agent is. Conversely, the tools tools we implement for the agent don't need to be as good at semantics.